Move MergeSelectionMasks to selection.py.
